refactor: updateMessage helper to public & private & uni data format
